+++ qpc_base.py
