+++ qpc_parser.py
@@ -6,8 +6,8 @@
 
 import os
 import hashlib
-import PyQPC_Base as base
-import PyQPC_Reader as reader
+import qpc_base as base
+import qpc_reader as reader
 
 
 class ProjectDefinition:
